requirements.txt
# GitHub Activities Tracker Dependencies

# GitHub API
PyGithub==1.58.2
requests==2.31.0

# Data Processing
# pandas and numpy are not currently used in the codebase
# but are kept for future data analysis features
# Both versions below are compatible with Python 3.12
pandas>=2.2.0,<2.3.0
numpy>=1.26.4,<2.0.0

# CLI Interface
click==8.1.3
rich==13.4.1

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Development
black==23.3.0
isort==5.12.0
flake8==6.0.0
